extracted feature-swim-lanes.module leaner AppModule: it depends on feature modules & BrowserModule as it should; other dependencies live in feature modules
