--- kanban-fire/src/app/app.component.css
@@ -1,34 +1,0 @@
-.content-wrapper {
-    max-width: 1400px;
-    margin: auto;
-}
-
-.container-wrapper {
-    display: flex;
-    justify-content: space-around;
-}
-
-.container {
-    width: 400px;
-    margin: 0 24px 24px 0;
-}
-
-.list {
-    border: solid 1px #ccc;
-    min-height: 60px;
-}
-
-.empty-label {
-    font-size: 2em;
-    padding-top: 10px;
-    text-align: center;
-    opacity: 0.2;
-}
-
-.cdk-drag-placeholder {
-    opacity: 0;
-}
-
-.cdk-drag-animating {
-    transition: transform 250ms;
-}
