+++ kanban-fire/src/app/app.component.spec.ts
@@ -1,32 +1,16 @@
 import { TestBed } from '@angular/core/testing';
 import { AppComponent } from './app.component';
-import { TaskComponent } from './task/task.component';
-
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
-import { DragDropModule } from '@angular/cdk/drag-drop';
-import { MatDialogModule } from '@angular/material/dialog';
-import { getFirestore, provideFirestore } from '@angular/fire/firestore';
-import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
-import { environment } from 'src/environments/environment';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { ToolbarModule } from './toolbar/toolbar.module';
+import { FeatureSwimLanesModule } from './feature-swim-lanes/feature-swim-lanes.module';
 
 
 describe('AppComponent', () => {
   beforeEach(() => TestBed.configureTestingModule({
     declarations: [
       AppComponent,
-      TaskComponent
     ],
     imports: [
-      DragDropModule,
-      MatDialogModule,
-      MatCardModule,
-      MatIconModule,
-      MatProgressSpinnerModule,
-      provideFirebaseApp(() => initializeApp(environment.firebase)),
-      provideFirestore(() => getFirestore()),
+      FeatureSwimLanesModule,
       ToolbarModule
     ]
   }));
@@ -35,18 +19,5 @@ describe('AppComponent', () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
     expect(app).toBeTruthy();
-  });
-
-  it(`should have as title 'kanban-fire'`, () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    const app = fixture.componentInstance;
-    expect(app.title).toEqual('kanban-fire');
-  });
-
-  it('should render title', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('.mat-toolbar span')?.textContent).toContain('Kanban Fire');
   });
 });
