+++ kanban-fire/src/app/app.module.ts
@@ -1,46 +1,21 @@
-import { BrowserModule } from '@angular/platform-browser';
-
-import { provideFirebaseApp, initializeApp } from '@angular/fire/app';
-import { getFirestore, provideFirestore } from '@angular/fire/firestore';
-import { MatButtonModule } from '@angular/material/button';
-import { MatDialogModule } from '@angular/material/dialog';
-import { DragDropModule } from '@angular/cdk/drag-drop';
-import { MatCardModule } from '@angular/material/card';
-import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner'; 
 import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { AppComponent } from './app.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { TaskComponent } from './task/task.component';
-import { TaskDialogComponent } from './task-dialog/task-dialog.component';
-import { FormsModule } from '@angular/forms';
-import { environment } from 'src/environments/environment';
 import { ToolbarModule } from './toolbar/toolbar.module';
+import { FeatureSwimLanesModule } from './feature-swim-lanes/feature-swim-lanes.module';
 
 @NgModule({
   declarations: [
     AppComponent,
-    TaskComponent,
-    TaskDialogComponent
   ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
-    DragDropModule,
-    FormsModule,
-    MatDialogModule,
-    MatButtonModule,
-    MatCardModule,
-    MatIconModule,
-    MatInputModule,
-    MatProgressSpinnerModule,
-    provideFirebaseApp(() => initializeApp(environment.firebase)),
-    provideFirestore(() => getFirestore()),
+    FeatureSwimLanesModule,
     ToolbarModule
   ],
-  providers: [],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
