+++ kanban-fire/src/app/feature-swim-lanes/task/task.component.ts
